requirements.txt
Django==4.2
YooMoney~=0.1.0
requests~=2.28.2
crispy-bootstrap5==0.7
django-crispy-forms==2.0
django-recaptcha==3.0.0
django-extensions==3.2.1
django-filter==23.1
psycopg2-binary==2.9.6
pillow==9.5.0